requirements.txt
reportlab
PyQt5
pdf2image
pytesseract
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
genai